Add IAsyncDisposable, IDisposable to PlotlyChart and PlotlyJsInterop
Fixes #483
